Move ZScatterer2Molecule to Molecule.h and Molecule.cpp, so it could be used from python
